+++ src/main/java/mchorse/mappet/EventHandler.java
@@ -1,7 +1,10 @@
 package mchorse.mappet;
 
+import com.google.common.collect.Lists;
+import com.google.common.collect.Sets;
 import mchorse.mappet.api.quests.Quest;
 import mchorse.mappet.api.quests.Quests;
+import mchorse.mappet.api.scripts.code.items.ScriptInventory;
 import mchorse.mappet.api.scripts.code.entities.ScriptEntity;
 import mchorse.mappet.api.scripts.code.items.ScriptItemStack;
 import mchorse.mappet.api.triggers.Trigger;
@@ -27,7 +30,10 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.EntityPlayerMP;
+import net.minecraft.inventory.Container;
 import net.minecraft.inventory.ContainerChest;
+import net.minecraft.inventory.ContainerPlayer;
+import net.minecraft.inventory.IInventory;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.DamageSource;
 import net.minecraft.util.EnumHand;
@@ -51,6 +57,7 @@ import net.minecraftforge.fml.common.gameevent.TickEvent;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
+import java.lang.reflect.Field;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -117,6 +124,18 @@ public class EventHandler
         return isMohist;
     }
 
+    public ArrayList<String> getIds()
+    {
+        ArrayList<String> ids = new ArrayList<String>();
+
+        for (IExecutable executable : this.executables)
+        {
+            ids.add(executable.getId());
+        }
+
+        return Lists.newArrayList(Sets.newLinkedHashSet(ids));
+    }
+
     public void addExecutables(List<IExecutable> executionForks)
     {
         this.executables.addAll(executionForks);
@@ -125,6 +144,15 @@ public class EventHandler
     public void addExecutable(IExecutable executable)
     {
         this.executables.add(executable);
+    }
+
+    public int removeExecutables(String id)
+    {
+        int size = this.executables.size();
+
+        this.executables.removeIf((e) -> e.getId().equals(id));
+
+        return size - this.executables.size();
     }
 
     public void reset()
@@ -220,11 +248,13 @@ public class EventHandler
             return;
         }
 
+        Container container = event.getContainer();
         DataContext context = new DataContext(event.getEntityPlayer());
+        IInventory inventory = null;
 
-        if (event.getContainer() instanceof ContainerChest)
+        if (container instanceof ContainerChest)
         {
-            ContainerChest chest = (ContainerChest) event.getContainer();
+            ContainerChest chest = (ContainerChest) container;
 
             if (chest.getLowerChestInventory() instanceof TileEntity)
             {
@@ -234,6 +264,36 @@ public class EventHandler
                 context.set("y", pos.getY());
                 context.set("z", pos.getZ());
             }
+
+            inventory = chest.getLowerChestInventory();
+        }
+        else if (container instanceof ContainerPlayer)
+        {
+            inventory = event.getEntityPlayer().inventory;
+        }
+        else
+        {
+            Field[] fields = container.getClass().getDeclaredFields();
+
+            for (Field field : fields)
+            {
+                if (field.getType().isAssignableFrom(IInventory.class))
+                {
+                    try
+                    {
+                        field.setAccessible(true);
+
+                        inventory = (IInventory) field.get(container);
+                    }
+                    catch (Exception e)
+                    {}
+                }
+            }
+        }
+
+        if (inventory != null)
+        {
+            context.getValues().put("inventory", new ScriptInventory(inventory));
         }
 
         Mappet.settings.playerCloseContainer.trigger(context);
@@ -322,6 +382,22 @@ public class EventHandler
             .set("hand", event.getHand() == EnumHand.MAIN_HAND ? "main" : "off");
 
         this.trigger(event, Mappet.settings.blockInteract, context);
+    }
+
+    @SubscribeEvent
+    public void onPlayerInteractWithEntity(PlayerInteractEvent.EntityInteract event)
+    {
+        EntityPlayer player = event.getEntityPlayer();
+
+        if (player.world.isRemote || Mappet.settings.playerEntityInteract.isEmpty())
+        {
+            return;
+        }
+
+        DataContext context = new DataContext(player, event.getTarget())
+            .set("hand", event.getHand() == EnumHand.MAIN_HAND ? "main" : "off");
+
+        this.trigger(event, Mappet.settings.playerEntityInteract, context);
     }
 
     /* Other cool stuff */
@@ -480,21 +556,6 @@ public class EventHandler
                 this.playersToCheck.add(player);
             }
         }
-    }
-
-    @SubscribeEvent
-    public void onPlayerInteractEntity(PlayerInteractEvent.EntityInteract event) {
-        EntityPlayer player = event.getEntityPlayer();
-
-        if (player.world.isRemote || Mappet.settings.playerEntityInteraction.isEmpty())
-        {
-            return;
-        }
-
-        DataContext context = new DataContext(player);
-        context.getValues().put("entity", ScriptEntity.create(event.getTarget()));
-
-        this.trigger(event, Mappet.settings.playerEntityInteraction, context);
     }
 
     @SubscribeEvent