+++ src/main/java/mchorse/mappet/api/npcs/NpcLexer.java
@@ -24,7 +24,8 @@ public class NpcLexer
         "speed", "can_swim", "immovable", "has_post", "post", "post_radius", "fallback", "patrol_circulate", "patrol", "follow",
         "faction", "morph", "sight_distance", "sight_radius", "drops", "xp",
         "look_at_player", "look_around", "wander", "flee",
-        "trigger_died", "trigger_damaged", "trigger_interact", "trigger_tick", "trigger_target", "trigger_initialize", "damage_delay"
+        "trigger_died", "trigger_damaged", "trigger_interact", "trigger_tick", "trigger_target", "trigger_initialize", "damage_delay",
+        "respawn", "respawn_delay", "respawn_on_coordinates", "respawn_pos_x", "respawn_pos_y", "respawn_pos_z", "respawn_save_uuid"
     );
 
     public String id;