+++ src/main/java/mchorse/mappet/commands/events/CommandEvent.java
@@ -7,6 +7,7 @@ public class CommandEvent extends MappetSubCommandBase
 {
     public CommandEvent()
     {
+        this.add(new CommandEventStop());
         this.add(new CommandEventTrigger());
     }